Add .gitignore. Almost finished softdelete (or call it "approval"). Added a mode "raw delete" with ensure this post won't be visible, even for admin. May be useful in the log, if you want to see which post have been raw-deleted. Still a little thing to do : When ensuring the post doesn't have anymore post (if ($post_data['first/last_post'])) it uses $post_data with deleted != 2. It should query like that, but should then add a foreach to remove deleted == 1 posts, to treat it differently. IE: the topic should be marked as deleted if there's no more post with deleted = 0 BUT it shouldn't deleted events as poll and stuff like that.

+++ includes/functions_post.php
@@ -881,10 +881,12 @@ function sync_topic_details($topic_id, $forum_id, $all_data_only = true, $skip_a
 	return;
 }
 
-/*
-* Delete a post/poll
-*/
-function delete_post($mode, &$post_data, &$message, &$meta, &$forum_id, &$topic_id, &$post_id)
+/**
+ * Delete a post/poll
+ *
+ * @todo remove it (use class_mcp instead)
+ */
+function delete_post($mode, &$post_data, &$message, &$meta, &$forum_id, &$topic_id, &$post_id, $method = class_topics::SOFT_DELETE)
 {
 	global $db, $cache, $config, $lang, $user;
 
@@ -898,61 +900,58 @@ function delete_post($mode, &$post_data, &$message, &$meta, &$forum_id, &$topic_
 		// MG Cash MOD For IP - END
 		include(IP_ROOT_PATH . 'includes/functions_search.' . PHP_EXT);
 
-		$sql = "UPDATE " . POSTS_TABLE . " SET deleted = 1, deleter_user_id = " . $user->data['user_id'] . ", deleter_username = '" . $db->sql_escape($user->data['username']) . "' post_id = $post_id";
+		$sql = "UPDATE " . POSTS_TABLE . " SET deleted = 1, deleter_user_id = " . $user->data['user_id'] . ", deleter_username = '" . $db->sql_escape($user->data['username']) . "', deleted_time = " . time() . " WHERE post_id = $post_id";
 		$db->sql_query($sql);
 
-		// Event Registration - BEGIN
-		/*
-		if ($post_data['first_post'])
+		if ($mode == class_topics::RAW_DELETE)
 		{
-			$sql = "DELETE FROM " . REGISTRATION_TABLE . " WHERE topic_id = $topic_id";
-			$db->sql_query($sql);
-
-			$sql = "DELETE FROM " . REGISTRATION_DESC_TABLE . " WHERE topic_id = $topic_id";
-			$db->sql_query($sql);
-		}
-		*/
-		// Event Registration - END
-
-//<!-- BEGIN Unread Post Information to Database Mod -->
-		$sql = "DELETE FROM " . UPI2DB_LAST_POSTS_TABLE . " WHERE post_id = $post_id";
-		$db->sql_query($sql);
-
-		$sql = "DELETE FROM " . UPI2DB_UNREAD_POSTS_TABLE . " WHERE post_id = $post_id";
-		$db->sql_query($sql);
-//<!-- END Unread Post Information to Database Mod -->
-
-		if ($post_data['last_post'])
-		{
+			// Event Registration - BEGIN
 			if ($post_data['first_post'])
 			{
-				$forum_update_sql .= ', forum_topics = forum_topics - 1';
-				$sql = "UPDATE " . TOPICS_TABLE . "
-					SET deleted = 1, deleter_user_id = " . $user->data['user_id'] . ", deleter_username = '" . $db->sql_escape($user->data['username']) . "'
-					WHERE topic_id = $topic_id";
-				$db->sql_query($sql);
-				$sql = "DELETE FROM " . TOPICS_TABLE . "
-					WHERE topic_moved_id = $topic_id";
+				$sql = "DELETE FROM " . REGISTRATION_TABLE . " WHERE topic_id = $topic_id";
 				$db->sql_query($sql);
 
-				/*
-				$sql = "DELETE FROM " . THANKS_TABLE . " WHERE topic_id = $topic_id";
+				$sql = "DELETE FROM " . REGISTRATION_DESC_TABLE . " WHERE topic_id = $topic_id";
 				$db->sql_query($sql);
-				*/
-
-				$sql = "DELETE FROM " . TOPICS_WATCH_TABLE . " WHERE topic_id = $topic_id";
-				$db->sql_query($sql);
-
-				/*
-				$sql = "DELETE FROM " . BOOKMARK_TABLE . " WHERE topic_id = $topic_id";
-				$db->sql_query($sql);
-				*/
 			}
+			// Event Registration - END
+
+	//<!-- BEGIN Unread Post Information to Database Mod -->
+			$sql = "DELETE FROM " . UPI2DB_LAST_POSTS_TABLE . " WHERE post_id = $post_id";
+			$db->sql_query($sql);
+
+			$sql = "DELETE FROM " . UPI2DB_UNREAD_POSTS_TABLE . " WHERE post_id = $post_id";
+			$db->sql_query($sql);
+	//<!-- END Unread Post Information to Database Mod -->
+
+			if ($post_data['last_post'])
+			{
+				if ($post_data['first_post'])
+				{
+					$forum_update_sql .= ', forum_topics = forum_topics - 1';
+					$sql = "UPDATE " . TOPICS_TABLE . "
+						SET deleted = 1, deleter_user_id = " . $user->data['user_id'] . ", deleter_username = '" . $db->sql_escape($user->data['username']) . "', deleted_time = " . time() . "
+						WHERE topic_id = $topic_id";
+					$db->sql_query($sql);
+					$sql = "DELETE FROM " . TOPICS_TABLE . "
+						WHERE topic_moved_id = $topic_id";
+					$db->sql_query($sql);
+
+					$sql = "DELETE FROM " . THANKS_TABLE . " WHERE topic_id = $topic_id";
+					$db->sql_query($sql);
+
+					$sql = "DELETE FROM " . TOPICS_WATCH_TABLE . " WHERE topic_id = $topic_id";
+					$db->sql_query($sql);
+
+					$sql = "DELETE FROM " . BOOKMARK_TABLE . " WHERE topic_id = $topic_id";
+					$db->sql_query($sql);
+				}
+			}
+			remove_search_post($post_id);
 		}
 
 		empty_cache_folders(POSTS_CACHE_FOLDER);
 		empty_cache_folders(FORUMS_CACHE_FOLDER);
-		remove_search_post($post_id);
 	}
 
 	if (($mode == 'poll_delete') || (($mode == 'delete') && $post_data['first_post'] && $post_data['last_post']) && $post_data['has_poll'] && $post_data['edit_poll'])
